+++ frontend/javascripts/test/backend-snapshot-tests/annotations.e2e.ts
@@ -11,8 +11,8 @@ import { createTreeMapFromTreeArray } from "viewer/model/reducers/skeletontracin
 import { diffTrees } from "viewer/model/sagas/skeletontracing_saga";
 import { getNullableSkeletonTracing } from "viewer/model/accessors/skeletontracing_accessor";
 import { getServerVolumeTracings } from "viewer/model/accessors/volumetracing_accessor";
-import { addVersionNumbers } from "viewer/model/sagas/save_saga";
-import * as UpdateActions from "viewer/model/sagas/update_actions";
+import { addVersionNumbers } from "viewer/model/sagas/saving/save_saga";
+import * as UpdateActions from "viewer/model/sagas/volume/update_actions";
 import * as api from "admin/rest_api";
 import generateDummyTrees from "viewer/model/helpers/generate_dummy_trees";
 import { describe, it, beforeAll, expect } from "vitest";
