+++ frontend/javascripts/test/helpers/saveHelpers.ts
@@ -1,5 +1,5 @@
 import type { TracingStats } from "viewer/model/accessors/annotation_accessor";
-import type { UpdateActionWithoutIsolationRequirement } from "viewer/model/sagas/update_actions";
+import type { UpdateActionWithoutIsolationRequirement } from "viewer/model/sagas/volume/update_actions";
 import type { SaveQueueEntry } from "viewer/store";
 import { idUserA } from "test/e2e-setup";
 import dummyUser from "test/fixtures/dummy_user";
