+++ frontend/javascripts/test/reducers/save_reducer.spec.ts
@@ -2,11 +2,11 @@ import { vi, describe, it, expect } from "vitest";
 import dummyUser from "test/fixtures/dummy_user";
 import type { WebknossosState } from "viewer/store";
 import { createSaveQueueFromUpdateActions } from "../helpers/saveHelpers";
-import type { UpdateActionWithoutIsolationRequirement } from "viewer/model/sagas/update_actions";
+import type { UpdateActionWithoutIsolationRequirement } from "viewer/model/sagas/volume/update_actions";
 
 import * as SaveActions from "viewer/model/actions/save_actions";
 import SaveReducer from "viewer/model/reducers/save_reducer";
-import { createEdge } from "viewer/model/sagas/update_actions";
+import { createEdge } from "viewer/model/sagas/volume/update_actions";
 import { TIMESTAMP } from "test/global_mocks";
 
 vi.mock("viewer/model/accessors/annotation_accessor", () => ({
