+++ frontend/javascripts/test/reducers/update_action_application/volume.spec.ts
@@ -16,7 +16,7 @@ import { diffVolumeTracing } from "viewer/model/sagas/volumetracing_saga";
 import type {
   ApplicableVolumeUpdateAction,
   UpdateActionWithoutIsolationRequirement,
-} from "viewer/model/sagas/update_actions";
+} from "viewer/model/sagas/volume/update_actions";
 import { combinedReducer, type WebknossosState } from "viewer/store";
 import { makeBasicGroupObject } from "viewer/view/right-border-tabs/trees_tab/tree_hierarchy_view_helpers";
 import { afterAll, describe, expect, test } from "vitest";
