+++ frontend/javascripts/viewer/model/reducers/update_action_application/bounding_box.ts
@@ -7,7 +7,7 @@ import type {
   DeleteUserBoundingBoxInVolumeTracingAction,
   UpdateUserBoundingBoxInSkeletonTracingAction,
   UpdateUserBoundingBoxInVolumeTracingAction,
-} from "viewer/model/sagas/update_actions";
+} from "viewer/model/sagas/volume/update_actions";
 import type {
   SkeletonTracing,
   UserBoundingBox,
