+++ frontend/javascripts/viewer/model/sagas/root_saga.ts
@@ -8,11 +8,11 @@ import listenToClipHistogramSaga from "viewer/model/sagas/clip_histogram_saga";
 import DatasetSagas from "viewer/model/sagas/dataset_saga";
 import type { Saga } from "viewer/model/sagas/effect-generators";
 import loadHistogramDataSaga from "viewer/model/sagas/load_histogram_data_saga";
-import MappingSaga from "viewer/model/sagas/mapping_saga";
+import MappingSaga from "viewer/model/sagas/volume/mapping_saga";
 import { watchDataRelevantChanges } from "viewer/model/sagas/prefetch_saga";
-import ProofreadSaga from "viewer/model/sagas/proofread_saga";
+import ProofreadSaga from "viewer/model/sagas/volume/proofread_saga";
 import ReadySagas from "viewer/model/sagas/ready_sagas";
-import SaveSagas, { toggleErrorHighlighting } from "viewer/model/sagas/save_saga";
+import SaveSagas, { toggleErrorHighlighting } from "viewer/model/sagas/saving/save_saga";
 import SettingsSaga from "viewer/model/sagas/settings_saga";
 import SkeletontracingSagas from "viewer/model/sagas/skeletontracing_saga";
 import watchTasksAsync, { warnAboutMagRestriction } from "viewer/model/sagas/task_saga";
