+++ frontend/javascripts/viewer/model/sagas/settings_saga.ts
@@ -13,7 +13,7 @@ import { type Saga, select, take } from "viewer/model/sagas/effect-generators";
 import {
   SETTINGS_MAX_RETRY_COUNT,
   SETTINGS_RETRY_DELAY,
-} from "viewer/model/sagas/save_saga_constants";
+} from "viewer/model/sagas/saving/save_saga_constants";
 import type { DatasetConfiguration, DatasetLayerConfiguration } from "viewer/store";
 import { Toolkit } from "../accessors/tool_accessor";
 import { ensureWkReady } from "./ready_sagas";
