+++ frontend/javascripts/viewer/model/sagas/volume/quick_select/quick_select_heuristic_saga.ts
