+++ frontend/javascripts/viewer/view/left-border-tabs/modals/add_volume_layer_modal.tsx
@@ -25,7 +25,7 @@ import {
   getVolumeTracingLayers,
 } from "viewer/model/accessors/volumetracing_accessor";
 import { pushSaveQueueTransactionIsolated } from "viewer/model/actions/save_actions";
-import { addLayerToAnnotation } from "viewer/model/sagas/update_actions";
+import { addLayerToAnnotation } from "viewer/model/sagas/volume/update_actions";
 import { Model, api } from "viewer/singletons";
 import Store, { type StoreAnnotation } from "viewer/store";
 import InputComponent from "viewer/view/components/input_component";
